mes: raise fuzzy match to 82.3% (#16042)
* mes: convert MakeAgbString inner tag dispatch to switch (81.3%)


* mes: express MakeAgbString nibble reads relative to cached next pointer (81.28%)


* mes: invert MakeAgbString FF check so char-copy lands after switch (81.56%)


* mes: convert addString font selection to switch (81.70%)


* mes: convert Next flag-replay dispatch to switch (81.74%)


* mes: pass &mText directly in Set loop instead of cast (81.94%)


* mes: convert Calc flag-replay dispatch to switch (82.06%)


* mes: convert useFlag dispatch to switch (82.19%)


* mes: convert drawTagString tag check to switch (82.25%)


* mes: tag 0x2F fallback is townName (82.32%)


* mes: GetMesNibbleValue sign-extends high nibble before OR (82.82%)


* mes: ReadTagByte sign-extends high nibble (82.83%)


* mes: branchMode unsigned char, execParam unsigned compares (83.00%)


* mes: advance in directly per-case, no read-pointer across switch (83.33%)


* mes: Next flag-replay switch includes case 3/4 like useFlag (83.37%)


* mes: signedness corrections in MakeAgbString tag handlers (83.67%)


* mes: signedness corrections in addString tag handlers (83.72%)


* mes: signedness corrections in Draw fade alpha (83.83%)


* mes: signed char read for fade nibble in Calc (83.86%)


* mes: switch + temp font in addString case 0x34 (84.0%)


* mes: split setup font temp in addString (84.1%)


* mes: mRubySpacing is float; split ruby advance product (84.2%)


* mes: switch for font selection in Draw loop (84.5%)


---------

+++ include/ffcc/mes.h
@@ -80,7 +80,7 @@ private:
     int mRubyHeight;            // 0x3D18
     int mRubyOffset;            // 0x3D1C
     float mRubyY;               // 0x3D20
-    int mRubySpacing;           // 0x3D24
+    float mRubySpacing;         // 0x3D24
     int mColor;                 // 0x3D28
     int mFontAlign;             // 0x3D2C
     int mFontCount;             // 0x3D30
